requirements.txt
# Core FastAPI
fastapi==0.103.2
uvicorn==0.23.2
websockets==15.0.1
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.10.1
pydantic-core==2.27.1
pydantic-settings==2.6.0

# Database
tortoise-orm==0.20.0
asyncpg==0.30.0
aiosqlite==0.17.0
aerich==0.7.2

# Authentication & Security
PyJWT>=2.0.0
python-jose==3.3.0
ecdsa==0.18.0

# Google Services
googlemaps==4.10.0
google-auth>=2.0.0
google-auth-httplib2>=0.1.0

# Firebase
firebase-admin==6.7.0
py-firebase-dynamic-links==1.0.1

# SMS
smsaero==1.1.0

# Phone numbers
phonenumbers==8.13.20

# HTTP clients
requests==2.31.0
httpx~=0.27.2

# Image processing
opencv-python==4.10.0.84

# PDF generation
reportlab==4.2.5

# Date/time
python-dateutil==2.9.0
pytz==2023.3.post1

# Testing
pytest~=8.3.3
pytest-asyncio~=0.23.0
asgi-lifespan~=2.1.0
beautifulsoup4~=4.13.3

# Logging
python-json-logger==2.0.7

# Other
typing-extensions~=4.12.2
